merdan

+++ hizmetler/views.py
@@ -80,20 +80,23 @@ def yorum(request, id):
     return redirect(reverse("hizmet:detay",kwargs={"id":id}))
 
 def artikle_likes(request, id):
-    user=request.user
-    article=Article.objects.get(id=id)
-    current_likes=article.like
-    liked=Likes.objects.filter(user=user,article=article).count()
-    if not liked:
-        liked=Likes.objects.create(user=user, article=article)
-        current_likes=current_likes+1
-    else:
-        liked=Likes.objects.filter(user=user, article=article).delete()
-        current_likes=current_likes-1
+    if request.user.is_authenticated:
+        user=request.user
+        article=Article.objects.get(id=id)
+        current_likes=article.like
+        liked=Likes.objects.filter(user=user,article=article).count()
+        if not liked:
+          liked=Likes.objects.create(user=user, article=article)
+          current_likes=current_likes+1
+        else:
+          liked=Likes.objects.filter(user=user, article=article).delete()
+          current_likes=current_likes-1
         
-    article.like=current_likes
-    print ("begenme", article.like)
-    article.save()
+        article.like=current_likes
+        article.save()
+    else:
+            messages.warning(request, "Begenebilmek için giriş yapmak gerekir")
+            return redirect('giris')
     
     return redirect(reverse('hizmet:detay',kwargs={"id":id}))
